requirements.txt
# Core dependencies
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn>=0.24.0
requests>=2.31.0

# RAG and AI dependencies
chromadb>=0.4.15
langchain>=0.0.350
groq>=0.4.1

# Data processing
python-dotenv>=1.0.0
pydantic>=2.4.0

# YouTube transcript processing
yt-dlp>=2023.10.13

# Utilities
glob2>=0.7
pathlib>=1.0.1

# Optional: For better logging and monitoring
rich>=13.6.0
tqdm>=4.66.0

# Web server
gunicorn>=21.2.0